Add preserveEdgeToEdge prop (#765)
## 📜 Description

Currently, even when `react-native-edge-to-edge` is installed, RNKC is
able to disable `edge-to-edge`. For example, if we set `enable={false}`
on `KeyboardProvider`, `disable` will be called on
`EdgeToEdgeReactViewGroup`, and `WindowCompat.setDecorFitsSystemWindows`
too.

This change detect if `react-native-edge-to-edge` is installed. If it
is, edge-to-edge is forced and cannot be disabled.

## 💡 Motivation and Context

It fixes an issue that breaks edge-to-edge (not only when enabled with
`react-native-edge-to-edge`, but in general).

Fixes
https://github.com/kirillzyusko/react-native-keyboard-controller/issues/592
(_Toggle enabled to false, observe that the NavigationBar reappears and
lifts the content._)

## 📢 Changelog

### Android

- Fixes issue with edge-to-edge when `KeyboardProvider` is used with
`enabled={false}`.

## 🤔 How Has This Been Tested?

- Using the example app (`react-native-edge-to-edge` installed). Before,
when switching on the "14. Enable / disable" screen, we were able to
disable edge-to-edge too, not only keyboard management. Now it stays
edge-to-edge.

## 📸 Screenshots (if appropriate):

![Screenshot 2025-01-19 at 22 59
31](https://github.com/user-attachments/assets/de7b9b9d-2b57-45ad-a960-f57ea3e0f43b)

## 📝 Checklist

- [x] CI successfully passed
- [x] I added new mocks and corresponding unit-tests if library API was
changed