feat: `insets` prop for `KeyboardToolbar` (#866)
## 📜 Description

Handle safe-area paddings in `KeyboardToolbar` component (landscape
mode).

## 💡 Motivation and Context

I had a lot of ideas how to fix this problem. Some of them:
- pass `insets` as prop;
- automatically grab insets from `useSafeAreaInsets` hook;
- render view wrapper (users can add their own SafeAreaView
implementation).

The **automatically grab insets from `useSafeAreaInsets` hook** idea has
failed, because in my case `react-native-safe-area-context` was
resolving to other packages (maybe it's the problem with package linking
in example app). But this is still not a perfect approach, because I'll
force users to install additional package.

The **render view wrapper (users can add their own SafeAreaView
implementation)** was perspective, but it would be very complex for end
users. Mostly because we have to provide additional props, such as
`backgroundColor` etc., so it would look like:

```tsx
<KeyboardToolbar
  wrapper={(props) => <SafeAreaView {...props} />}
/>
```

And there is still a lot of open topics on how to handle many aspects -
we would have to write conditional code when wrapper is not provided and
render `View`.

So in the end I decided to add `insets` property - and it would be up to
lib users to provide it or not. Maybe not ideal solution, but definitely
better than having an open issues for 9+ months 🙃

Closes
https://github.com/kirillzyusko/react-native-keyboard-controller/issues/572

## 📢 Changelog

<!-- High level overview of important changes -->
<!-- For example: fixed status bar manipulation; added new types
declarations; -->
<!-- If your changes don't affect one of platform/language below - then
remove this platform/language -->

### JS

- added new `insets` property.

## 🤔 How Has This Been Tested?

Tested manually on iPhone 15 (iOS 17.5), Fabric.

## 📸 Screenshots (if appropriate):

|Before|After|
|------|-----|

|![image](https://github.com/user-attachments/assets/a90955a5-c33b-44d9-80fb-e4e7e88bc735)|![image](https://github.com/user-attachments/assets/c64a54df-f0e9-4f5a-b62b-26b2f679134b)|

## 📝 Checklist

- [x] CI successfully passed
- [x] I added new mocks and corresponding unit-tests if library API was
changed