fix: wrong `useWindowDimensions` on iPad (#957)
## 📜 Description

Add support for `windowDidResize` event on iOS.

## 💡 Motivation and Context

In
https://github.com/kirillzyusko/react-native-keyboard-controller/pull/948
I changed implementation and added subscription to `screen` values
instead of `window`.

While it fixes some issues it doesn't work on iPad 🤯 This is very sad
that such basic functionality can not be handled by RN out-of-the-box.
However we already use own `useWindowDimensions` hook on Android. So in
this PR I decided to replicate that behavior on iOS.

`react-native-safe-area-context` doesn't have that problem, but
`react-native-safe-area-context` handles that event in `layoutSubviews`.
After thinking a while I decided to use the same approach and emit event
`windowDidResize`. And this approach actually works 🙂 I also added a
check to not emit events too frequently if dimensions weren't changed.

The PR
https://github.com/kirillzyusko/react-native-keyboard-controller/pull/956
also aims to solve that problem, but `useWindowDimensions` is used by
other components, such as
`KeyboardAvoidingView`/`KeyboardAwareScrollView`, so the problem should
be fixed in hook itself.

Closes
https://github.com/kirillzyusko/react-native-keyboard-controller/issues/955

## 📢 Changelog

<!-- High level overview of important changes -->
<!-- For example: fixed status bar manipulation; added new types
declarations; -->
<!-- If your changes don't affect one of platform/language below - then
remove this platform/language -->

### iOS

- emit `windowDidResize` event with device screen dimensions;

## 🤔 How Has This Been Tested?

Tested manually in reproduction example + by e2e tests.

## 📸 Screenshots (if appropriate):


![image](https://github.com/user-attachments/assets/d2379f1d-679d-4314-b6d8-840d667c65b6)

## 📝 Checklist

- [x] CI successfully passed
- [x] I added new mocks and corresponding unit-tests if library API was
changed